sqlite, sqliteh: move Tracer interface
So we can use the Tracer without the driver.
